Fix: Better Feedback for crashes when calling APIs from Background Threads
This Boeing 747 of a library is very analog. We've been extending it, granted, but to keep it "within the 747 family" we're trying to match it as much as possible. An all-encompassing rewrite is not in the cards, currently. But the structure of the library expects to always be called externally from the Main Thread, so, in the past we've thrown errors / crashed on purpose to alert developers not to do this. It is also in bright red "Caution" in the README to do this. Anyway, we've changed the structure of the code around so the stack trace makes it a lot more explicit as to what the underlying issue is.

This is in reference to GitHub Issue https://github.com/NordicSemiconductor/IOS-nRF-Connect-Device-Manager/issues/383