Fix: McuMgrBleTransport now checks if current MTU is valid after Connection (Result)Lock
This explains why making independent API calls, such as an echo or a single mcuMgrParameters before a full-DFU looked smoother, it's because there was a delay that allowed the transport to get a valid MTU value after an active connection was established. Now we check at the lowest level; I attempted to throw an "invalid MTU" value but I think it caused too many issues. Doing it here is sneaky, but it's also the Transport's responsibility? Anyway, things should work better now.