Fix: Wait for BaseViewController's deviceStatusInfo Requests
Previously, BaseViewController was firing simultaneous requests for mcumgr parameters and others in parallel to an app's operation, such as start DFU. This caused locking issues in the sense that, the transport was reused for DFU and BaseViewController requests, and the supposedly random sequence number was not so random so, different requests happened to land on the same internal 'lock' causing issues. We fixed that by moving the BaseVC to its own private transport, but we then learned that many firmware(s) will not accept these simultaneous requests, so we modified the app to "wait" for BaseVC's requests before making a big operation such as DFU or FileManager.