Observability (aka 'nRF Memfault Library Functionality') added to nRF Connect Device Manager
There's less shine in this instance. The UI is more brunt force, not showing all the chunks sent and their data not allowing to per-chunk retries when needed. But, in the perfect-use-case it should be working. Overall, both OTA and Observability require polishing work but, basic functionality is now present. The Observability is the nRF Memfault Library code ported over with some changes, though it is mostly the same code as over there. Adapted to the newer iOS-BLE-Library.