Fix: Don't relaunch BaseViewController's Observability Task on "isDeviceStatusReady" callback
Basically, if we connect to a Device by any means, and then tap "Check for Updates", the Observability Task will relaunch and reconnect. There's no harm, but there's no need for the app to do that, either. The quick drawback is that Observability's BLE communications are now "in parallel" with SMP, which is unwanted as far as I know. But, tradeoffs. Specially against time.