Fix: Detect "Firmware Already up to date" scenario
We now show a specific alert to the user if there is no update available via nRF Cloud OTA. The implementation is a bit tricky because our current "Network" implementation does not return or allow us to see the status code, it just forwards us back the response Data bytes, which in this care they're empty. The quickest solution was to hijack the catch error path to detect this use case.