Restored LICENSE in main README file
Because we moved it in the shuffle / reorg, of which we now have to do a new round of, the GitHub indicator linked from the README is broken. This restores it. I opted to copy it, so the McuMgrCocoapod retains its own copy as well, which is why it was moved there.