Migrated API to gather DeviceInfo and Project Tokens from iOSOtaLibrary into nRF CDM code
Proper procedure to obtain the two components / tokens to perform an OTARequest, has not been decided upon. Initially it's being set up as a combination of values from Device Information Service, standard GATT, as well as the existing MDS (Memfault Diagonostic) Service. Since this is not stable footing, and is not guaranteed to be a forward-looking API, before publishing the new iOSOtaLibrary component, we're moving it away. But it's still Open Source and accessible, of course.