Fix: Detect Unreachable Network
This is not the best, because, our "Network" code implementation is duplicated here and in iOS-Common. I think now we have some more bandwidth to bring the minimum version of iOS-Common down and be able to solve this for the future. But for now, this fixes issue https://github.com/NordicSemiconductor/IOS-nRF-Connect-Device-Manager/issues/447