Fix for unrecognised McuMgrReturn Codes
Our listing of return codes 'ends' at 13. But, Group Return Codes extend, for example in the case of Image Group, past 23 or 24. So, because there is no direct matching in McuMgrReturn Code, we match it to "ok" by default, and then it all falls through. So, we need to differentiate the two, and we cannot match a single enum case, unrecognised, to multiple values I think. So we fixed it on the implementation side with @returnCodeValue and that's the redirect we do with group errors. It is not perfect, but it is better than before.