Change: In Test & Confirm Mode, precede a "List" Command to Confirm Stage to match Image, Slot combinations
If we upload an image to Image 0, Slot 1, in Test & Confirm we'll then send "Test" command, verify it's Tested, then Reset, and after Reset, send a "Confirm" command. After confirming it has been confirmed, we call success. However, an image uploaded to 0,1 can (and is expected in 52 with no DirectXIP) to be swapped to 0,0 and marked as Active. But then, we are no longer confirming image 0, slot 1, but image 0 and slot 0. So, we perform a List command, match against our @images array, and then continue. In the logs, we will then be able to correctly log "confirming image 0, slot 0'" for example. It is all very confusing because @images before reset is the "target result" and not the current status, and we are changing this for the specifics of Test&Confirm, but anyway, I can't think of anything better to resolve this and future issues due to the inherent Reset swap at the moment. I could be wrong, of course.

Prompted by: Promoted by: https://github.com/NordicSemiconductor/IOS-nRF-Connect-Device-Manager/issues/471