Implemented Auto-Retry when Network is Unavailable
I know this is not perfect, because neither the @networkBusy flag nor this single Timer would work well with multiple connections to multiple devices. Which, in other parts of the code, we've tried to keep in mind as a possibility. But, right now, my brain can't deal with that. But I can track this one through and implement it. Not perfect, but perhaps good enough for an initial release version.