Efficient Writes of ObservabilityState to disk
It's probably overkill, perhaps it is. But, it's a good thing to not write to flash if we can avoid it. So here, all we want to do is "accumulate" writes into one, if we can do so. That's all it is. If we receive 12 chunks at once, we try to write once that we now have 12 chunks. And when we manage to upload them, something similar.