Fix for "bad-MTU" correcting code in McuMgrBleTransport
It's possible that the MTU might need to be set lower than maximum. For example (as in testing currently), if the user has a Reassembly Buffer Size smaller than MTU size. This will trip the update code because we will force upper layers to send maximum MTU size packets, when the target device doesn't want to. In theory, MTU should be set to match buffer size, but anyways. I think this was working before, but our speed improvements broke this.