Added safeguards for minimum pipelineDepth value
This affects both "regular" DFU as well as the FileManager. Essentially, the pipeline depth / number of buffers is a parallel queue of sequence numbers, even though we cannot send them in parallel / interleaved because then the target device cannot make sense of the Data being sent. But, the point is, in a way it sets the number of "tasks" to send or add to a queue in parallel. We were too busy checking we are doing the translation between "number of buffers" and pipeline depth correctly, since UI and the code interpret the same value in different ways (essentially, a -1 arithmetic operation) that we did not bother to check whether the pipelineDepth is set to a value that soft-locks the code... because nothing is sent? If we set the pipeline to zero, then nothing is enqueued to be sent. Essentially freezing everything.

Prompted by: https://github.com/NordicSemiconductor/IOS-nRF-Connect-Device-Manager/issues/497