Fix for stalled FileSystem uploads when Buffer Size is smaller than MTU
We solved this (again) for DFU, so it was trivial to check for FS, and it turned out it was broken as well, which was surprising because the first thing we checked was the source code, and the source code seemed to account for this scenario, but it was done so after the first packet upload, which is already too late for this particular scenario. We broke it ourselves, if I had to guess.