Buttonless DFU for NCS Bare Metal
Added the concept of "modes" to McuMgrTransport, which is a bit... I tried, or did my best, to not add any hacks. The transport takes care of the peripheral, and it still does. We just tried to add the capability for the same transport to talk to different peripherals, although in practice it's the same device. This is the less hacky way we could find. The transport also has never been the one with the responsibility yo "find" a device, so we added a new helper type to FirmwareUpgradeManager to handle that.

Note that this implementation targets sample code that is subject to change with no prior warning.