Refactored FirmwareUpgradeImage into its own file
Like, the Manager file was close to 2000 lines? In the early days I just wanted to survive when given tasks to do around here, but at some point we have to clean this up. If only for my own sanity since I have to write more code and make it grow again.